Limit the responsibilities of TaggingUpdateForm
TaggingUpdateForm is now only for capturing user input from the form.

ContentItemLinks replaces the old LinkSet model. ContentItem.link_set now
returns a ContentItemsLinks object, and the form can be initialised using that.

This is in preparation for adding related items to the form. The form will show
base paths, but these need to be mapped to/from content ids when communicating
with the publishing api.