Fetch expanded links instead of links
We now fetch the expanded links from the publishing api instead of just
arrays of content ids. We'll use this in the tagging update form so that
we can work with base paths instead of content ids.

ContentItemExpandedLinks is only responsible for retrieving links.
We're deliberately handling submitting links back to the publishing api
using `patch_links` separately, because the representation of the
links is different.