Check blacklisted types on #update
Don't rely on the presence of `nil` (i.e. the field was
not submitted in the form) to determine whether the tag type is blacklisted.

Do the same blacklist check on the backend so that regardless of what the user
submits the backend will do the right thing for the format.

A form field value should never be null normally, and if it is, we can treat
it as empty.