Refactor tagging form handling
This refactors the way we create the tagging form and handle the user
input.

It attempts to improve the readability of the controller and
`TaggingUpdateForm`.

The basic story is:

- The show action in the controller instantiates a `TaggingUpdateForm`
object. This is an ActiveModel-compliant object and it's only
responsibility is to allow us to build the form.
- The `update` action uses the `TaggingUpdatePublisher` class, which
has the task of parsing the user input and sending it to the
publishing-api.

This also makes sure that we use just one instance variable in the
controllers, making us a little more Sandi-compliant
(https://robots.thoughtbot.com/sandi-metz-rules-for-developers).