Reorganise everything into namespaces
Big commit, sorry.

This organises most of the code into one of three namespaces: Tagging,
BulkTagging and Taxonomy.

The goal of this is make the application easier to understand to
newcomers and to avoid it becoming a big ball of code. The three
features of this application are pretty well-defined, and should not
have to talk to each other.

Some notes:

- I considered inverting the pattern/namespace (so that it reads
`app/taxonomy/models`, `app/bulk_tagging/services`). That seems a
little outside of the mainstream, but might be worth talking about
later.
- There are a couple of classes that are used by multiple modules. I've
kept these at the root level. The rules of this separation are that no
module can talk to each other, but that each module can talk to the top
level. The top level should also not talk to the modules, but that
requires refactoring outside the scope of this PR.