Rake task for removing unused taxons
Unpublish old taxon content items that are not linked to the current
education taxonomy.

Anything that is not reachable from the top level of the education
taxonomy is no longer needed and clutters up content tagger.

None of these have been visible to users so don't bother with redirects.

There is a separate task for just printing the orphaned taxons so we
can see what will be removed.