Rename education base path to /education
Top level themes are the root taxon of a tree. Every other taxon in the tree
has that initial URL as a prefix.

Because of https://github.com/alphagov/content-tagger/pull/268, the taxon
"Primary curriculum, key stage 2" would have a URL like this:
https://www.gov.uk/education-training-and-skills/primary-curriculum-key-stage-2

In https://github.com/alphagov/content-tagger/pull/275, we removed everything
that is not under the education theme.

This commit renames the prefix to /education.

Some taxons are not valid at the moment (due to missing descriptions) but the
rake task ignores this so we can update the base paths independently.

For now if anything fails to be updated, skip it.

We'll remove /alpha-taxonomy, and /education-training-and-skills in future
commits, so that the validation still works if we can't migrate all taxons at
once.

https://trello.com/c/CbJ3kzQC/439-rename-the-base-path-of-education-training-and-skills