Store errors for failed publishing
When we import a spreadsheet, we create TagMapping records for every tag
we are bulk loading. We queue sidekiq jobs to publish the links, but if
they fail, the records are left as "Ready to Tag", and the page seems to
hang forever.

In this case, we should admit the thing has errored so it's clear that
something has gone wrong.

We need to use `with_connection` because the block executes outside of
ActiveRecord middleware. See the configuration section of
https://github.com/mperham/sidekiq/wiki/Error-Handling