Use taxonomy helpers gem to parse child taxons
The publishing API can now return all the descendents of a taxon
in one expanded links response.

The `govuk_taxonomy_helpers` gem can parse this response, and the `TreeNode`
class has been extracted to `LinkedContentItem` in that gem, so we can reuse
it in other apps.

We can't completely remove the taxonomy parsing code from content tagger,
because it still supports multiple parents for a taxon, which the gem doesn't.
We decided to keep this functionality for now, even though we're not using it
with the education taxonomy.

The `ExpandedTaxonomy` class handles this by separating the taxonomy into two
hierarchies, the "child expansion", rooted at the current node, and the "parent
expansion", also rooted at the current node, but inverted.

So for example if you are looking at "Jaffa Cakes", the `ExpandedLinks`
structure might look like

 Biscuits (depth 1)   Cakes (depth 1)
        \               /
         \             /
          \           /
       Jaffa Cakes (depth 0)
         /            \
        /              \
 12 pack (depth 1)   36 pack (depth 1)

This change swaps out the child expansion, while keeping the parent expansion
the same.