Use `get_content_items` to populate linkables
Previously we were using `get_linkables` to fetch all types of links
from the Publishing API to populate dropdowns and allow tagging of
pages.

These links include:
- taxons
- organisations
- topics
- mainstream browse pages
- needs

These API calls have become increasingly slow which means that in most
environments the requests to this particular page timeout.

This commit replaces calls to `get_linkables` to `get_content_items` in
the hope to speed the load page of that tagging page and prevent
timeouts in all environments.