Add taxon undelete functionality
This commit adds the ability to undelete previously deleted taxons. Deleted taxons are listed on a separate “trash” page, from where they can be restored.

The taxon deletion process is also changed as follows:

* The delete confirmation page now makes it clear that deleted taxons can be restored
* The wording of the “delete” button is now less scary, since deletions can be reversed

Users attempting to create a new taxon with the same slug as an existing one (live or deleted) now reveive a more specific error message to this effect.

As a bonus, error messages when creating/editing taxons are now displayed correctly, and the values of form fields are not lost if there is an error when attempting to publish the taxon.