Rake task to upload the parenting and childcare taxonomy
This rake task takes a CSV file containing taxon titles, descriptions
and parents, and creates them in content tagger.

The reason for this task is so we can upload the new taxonomy to staging
and perform a data dump for the prototype. It will also let us bulk tag
all pieces of content related to the parenting and childcare branch, so
they are available in search. This way, we can point the prototype to
staging and be able to show both scoped search results and related
links.

**NOTE**: this rake task is temporary and only needed until we can
create draft taxons live. Since the prototype will be tested before
that's available, we need it as a temporary step.