Change workflow to publish published taxons immediately
When taxons were updated previously they were not immediately published.
This meant that if a user edited a published taxon it would appear
both in the published list and the draft list. Published taxons are now
published immediately when changes are made to them so that they only
appear in the published list.

Trello: https://trello.com/c/8YH5l9kF/517-epic-implement-the-draft-workflow-for-taxonomy