Inline govuk_taxonomy_helpers
This inlines the relevant classes from the
https://github.com/alphagov/govuk_taxonomy_helpers gem.

The taxonomy gem was part of an effort to avoid duplication between
applications dealing with the taxonomy. Unfortunately we ran out of
time last quarter, which left us with an API that was still in flux
(alphagov/govuk_taxonomy_helpers#11).

To avoid confusion, I think it's best to inline the code in the places
where we use the taxonomy helpers. We're going to be doing work around
the taxonomy in the next couple of months and I think it will be easier
to extract code than trying to nail the API right now.

Trello: https://trello.com/c/FVfbiKxn