Add temporary rake task to unpublish /api
We are retiring the Content API. The application used to publish a
special route: `/api`. Since the application has now been retired, we
need to unpublish that content item.

This commit adds a temporary rake task that will unpublish the enpoint.

**Note**:

Ideally, this should have been done in the app itself. When I realised I
needed to unpublish it, the app was no longer deployable, which is why
I'm creating this temporary rake task in `content-tagger`.

After the rake task runs in production, we should delete it.