Commonize sortable form element for taxon tagging
We have two extremely similar sortable fields, used for tagging related
content to content items. We will soon be adding a third of these fields
(for external related links), so in preparation for this, this change
separates the repeated HTML from the content of the form elements, so
that it can be reused without copying and pasting.