Add a task to export a taxonomy to a JSON array
The navigation prototype needs to display the Parenting & Childcare
taxonomy, but this taxonomy is still only available in the draft
content store. Since the draft content store is in a VPN that the
prototype cannot access, we need to manually export the taxonomy so
that we can manually mock out a content store response.