Publish world taxonomy leaf first
We want to make sure that world taxonomies are published leaf first.
This means that when /world/<country> is published, the child taxons
already exist in publishing api and the links hash will be immediately
populated. This way the pages are immediately published with all their
child taxons in place. If we allow a random order of publish, we risk
the /world/<country> publishing before the children which means that
when a child of the <country> taxon publishes it needs to republish the
<country> (parent) before it shows up in the parent's links. This risks
the page for /world/<country> displaying without a child until
dependency resolution has run through (and this could take a while
depending on the sidekiq queues).