Allow UpdateTaxon to handle nil for associated taxons
When attempting to programmatically update taxons, an error is
thrown for `empty?` not being a method of `nil`. When loading
a taxon by content_id, the associated_taxons are returned as
nil. I think this error has not previously been seen since the
form will return an empty selection to an empty array.
It's still useful to be able to update taxons programmatically
though, so this change enables that.