Mainstream Browse taxonomy import tools
Provides a framework for creating the modern GOV.UK taxonomy from the
Mainstream Browse taxonomy.

The UI provides two rake tasks:

 - rake taxonomy:mainstream_browse
 - rake taxonomy:load

The first of these will construct a taxonomy data structure and save it
to tmp/msbp.yml

The second command will load this YAML file and persist it to the
publishing platform via publishing-api endpoints.