Show modal when clicking links in project show
Clicking on a link on the project show page opens up a modal with the contents
of the link embedded in an iframe.

The contents in the iframe is proxied through the iframe enabling proxy to work
around the X-Frame-Options setting in the header.

Right clicking opens the page (not proxied) in a new tab/window.