DRY up the root taxon contstant
The HOMEPAGE_CONTENT_ID taxon was duplicated a bunch of times in
different classes. This adds a GovukTaxonomy module with the constant
and makes everything use that. I've renamed it ROOT_CONTENT_ID to
communicate that it's the root of the taxonomy. We could just as easily
use another content item for this.