Redesign the management of root taxons
Move the previous index page for root taxons to be an explicit edit
page, and change the index page to a show page with a table. This show
page is now used to link out to the pages for the individual root
taxons.

This will allow users to navigate to the show pages for root taxons.