Update project filters to new states
This changes the tagging participant's workflow to show 'incomplete'
content items as the default view in the project page. Once a content
item has been flagged or marked as done, they will be filtered out from
the default list.

The filters have been updated to 'to do' (default), 'flagged' and 'done'.