Refactor the TaggingHistory::IndexPage class
Rename to LinkChanges, as this is now used by both the index and show
pages. Split out the handling of changes and users in to separate
methods. Handle the link_types parameter in the class, and remove it
from the controller. Don't reverse the results, as this is unnecessary
now that the Publishing API order has changed.