Establish order on projects index page
It was observed that the order of projects displayed on the index page
is seemingly random, and was re-shuffled when an update was made to a
project content item.

In this commit, we have defined the order in which projects should be
displayed: We are grouping by taxanomy_branch and displaying it's
title, but we do not have access to the title in the database, only the
taxanomy_branch UUID. This is used to define the order of the grouping,
and then order individual projects by names.