Recalculate waypoint trigger points after modifiying DOM
We are initialising the select2 boxes using the Waypoints library,
however, this library works solely on scroll events.

In a normal workflow, our users do not scroll down the page, but
instead mark the content as "Done" or "Flagged". These actions will
remove the elements from the page (which then brings the element below
up into view) while never triggering a scroll event.

We resolve this by refreshing the Waypoint library when changes to the
DOM have completed.