Refactor the ProxyIframe module
Switch to using events on the modal, rather than adding on click
handlers to the title links, as this is neater, and allows for hooking
in to when it begins to open, as well as when it is closed.

Locate the h4 and iframe elements by looking through the DOM, rather
than using ids, as this is easier to follow.

When clearing the iframe, set the URL to about:blank, rather than
deleting the body. This means that it works with other types of
content, e.g. PDFs.

Remove the now redundant ids from the modal elements.