Make 'suggested tags' field optional
Makes the 'suggested tags' field when flagging a ProjectContentItem as 'there's not relevant topic' optional. This was enforced via JavaScript so we just remove the relevant lines.

We also don't want flagged ContentItems that have no suggested taxons to
show up on the index page. This commit updates the FlaggedContentQuery
class to not return those content items. We still need to return all
content items that are flagged with "needs_help".

Since this makes the FlaggedContentQuery more complex, we add a test
that documents this behaviour.