Allow TaxonomySize #max_size to handle empty array
It's possible for `TaxonomySize#add_children_recursively` to return
an empty array, as was the case with the Corporate Information list.
Previously this would error as it would eventually try to run the
`max` method on `nil`. Instead the base case should trigger for empty
arrays as well.

We add a test for the `max_size` method to cover this edge case.