Change legacy policy taxon base_paths
The same policy can be related to multiple PolicyAreas. Previously the importer would not generate unique taxon base paths for the taxons, so the PublishingAPI would return GdsApi::HTTPUnprocessableEntity for when a base_path was already taken by a previously created taxon. This commit updates the PolicyTaxonomy class to generate unique taxon names for each policy area branch.