Auto-prefix taxon path based on the parent
We have removed the Theme concept from content-tagger as it's a
restriction that was hindering the Taxon creation process. Before, it
was mandatory to assign a Theme to a Taxon, which was primarily used to
defined path prefix (e.g. /education). This URL structure was defined
earlier this year [1].

Now we have removed the Theme model and requirement, the URL restriction
is dictated by the selected parent taxon:

- For root taxons (i.e. a taxon with no parent taxon), the user is free
  to define the slug.
- For non-root taxons, the path will be prefixed by the selected parent
  taxon, e.g. if a parent taxon is selected that is from the Education
  branch, the path will be prefixed with "/education".

[1] https://gov-uk.atlassian.net/wiki/spaces/FS/pages/127664172/URL+structure+for+the+new+navigation+pages