Set visible_to_department_editors to true on publish
Add a `PublishService` class to encapsulate the logic for toggling the `visible_to_department_editors` field. This value needs to be set to true for any taxons that are published. This commit goes hand in hand with code changes to NavigationHelpers to enable showing/hiding taxons based on the `visible_to_department_editors` value.