Simplify path slug validation
A valid path would be /highest-level-taxon-name/taxon-name, and as we
are automatically prefixing the /highest-level-taxon-name, the user is
restricted to the last part of the path, and only to a single level.

We change the taxon form field to not require a leading slash as it's an
unnecessary requirement for the user when they can only define paths to
one level anyway.