Show a friendly error message on publish failure
When trying to publish a taxon with the Publishing API sometimes
this can fail. Rather than showing the user a 500 error, it seems
like a good idea to show a slightly more useful error message
while also recording the problem in Sentry for us to investigate.

This was found since Content Tagger relies on Publishing API to
perform validation of base paths, and we will shortly be adding
checks to make sure they are unique regardless of letter case
which may affect users of the application.