Add metrics for content coverage to Content Tagger
This is a lift and shift from the govuk-tagging-monitor. The Topic
Taxonomy is less about navigation now, and a more general tool to help
users find what they are looking for on GOV.UK.

Moving the metrics for content coverage in to Content Tagger will
allow for tracking these metrics in each environment, rather than just
production, and also enable less duplicated code between the
govuk-tagging-monitor and Content Tagger.