Update query used for organisation tagging progress
Using the reject_taxons field to calculate the number of tagged content
items did not correctly represent the Topic Taxonomy. Now that we have
taxons imported from policies, mainstream browse, etc. there are a
number of content items that are not tagged to the Topic Taxonomy.
However, they are "tagged" and so the reject_taxons field is not fit for
purpose anymore.