Remove "Choose one..." placeholder from taxon typeahead
When the placeholder text shows, this means that the content item has no
taggings at all.

When the placeholder text does not show, but the typeahead box shows as
empty, this means that the content item has taggings but from other
branches.

From the perspective of the project's view, these two circumstances are
the same: the content item doesn't have any tags applied to it that are
from the project's branch.

It is confusing to the user as to why some boxes show the placeholder,
and others do not but by removing the placeholder text, we should be
able to remove this confusion.