Rewrite the TaxonomySize class and related presenter
Change the name to TaxonsWithContentCount, as this hopefully is a
little clearer that TaxonomySize.

Change this to use the GovukTaxonomyHelpers::LinkedContentItem to do
the tree traversal, to reduce duplication.

The aim here was to make this work for the homepage, as well as other
taxons, but I'd rather not add special cases in here, and instead add
support to the GovukTaxonomyHelpers.