Remove Bootstrap modal fade animation for flags modal
This is one of a series of changes to resolve our intermittently failing
feature spec (flagging_a_content_item_spec.rb).

There is an issue in some Capybara drivers, such as poltergeist being
the one that we are using, when it comes to CSS animations. There is a
chance that the element is moving when the driver is calculating the
element's position. The driver would normally then click on that
coordinate to trigger the click event, similar to what the end-user
would be doing. However, a race condition exists that the element
position may have changed from the CSS animation from the first
calculation.

As a quick solution, we have removed the fade class from the Bootstrap
modal to stop animations occuring.

An alternative solution could be to use .trigger('click') on the
element, but this does not replicate a user clicking on something on the
screen. Another alternative could be to disable CSS animations (with
CSS) for the test environment only.

For the curious, you can read more about the problem at this poltergeist
issue: https://github.com/teampoltergeist/poltergeist/issues/530