Wait until HTML is rendered before initialising RadioToggle module
Another problem we were having with the intermittent feature spec
(flagging_a_content_item_spec.rb), was how the flag HTML content was
being injected after clicking "Flag for review".

There was a small chance that the start function was being invoked
before the HTML was fully rendered.

We can work around this by attaching the Promise behaviour to the html
function, and then waiting for that promise to be resolved.