Display conflicting base path error message on publish action
Currently a "generic error" message is displayed when trying to publish
a draft content item that has a conflicting base path in the published
state.

We can perform a lookup on the bae path if a 422 error response is
returned by the publishing API to determine if this specific error case
is try, and provide a more detailed error message to the user interface.