Prevent invalid content items from being imported
When creating a project via the spreadsheet upload, we create content
items with incomplete data. The spreadsheet only contains the base_path,
title and description, but during the import, we queue asynchronous jobs
to fetch and update all the recently created content items using the
Publishing API to backfill the content_id field.

For the most case, this work fine, but it does not perform as expected
when importing content that already exists in Content Tagger. In fact,
the user does not know that the project may have been partially
imported, or not at all.

This commit changes the spreadsheet upload in a number of ways:

1. Moves the content_id lookup to be synchronous and using a single
batch /lookup-by-base-path request instead of multiple singular
requests.

2. Surfaces the problematic content items when attempting to creating a
project with content items that may exist in other projects already. We
introduce a Rails uniqueness validation to check that the content_id
could be taken, and bubble up a list of 'invalid content items' to the
view when the creation fails.

Before, we were reliant on a database unique index to prevent duplicate
content items from being created, but this was not easily detected since
they were being raised in the asynchronous jobs.