Fix base path preview when "Root of the taxonomy" selected as parent
The logic that exists for the base path preview assumed that a parent
taxon selected would have a path_prefix, and we use this assumption to
form the base path preview.

However, if the "Root of the taxonomy" taxon was selected as the parent,
the base path preview was producing a path such as "/undefined/taxon".
This is because the top of the tree has a base path of "/".

We have therefore added an exception for this particular taxon, so that
the base path preview only uses the path_slug to produce the preview
when the "Root of the taxonomy" taxon is selected as the parent.