Change behaviour for users with no permissions
Signon has a concept of access (which means the signin permission),
and other permissions (which are "GDS Editor" and "Tagathon
participant" in the case of Content Tagger).

Previously, just giving a user access to Content Tagger in Signon,
would not have given them any access to the app. This commit changes
that behaviour, and gives them access to see the taxonomy, but not
make any changes through Content Tagger.

The aim here is to neaten up the user experience. Now that it's
possible for a user of Content Tagger to have access to see the
taxonomy, but not to change it, this makes a good default experience
for users without other permissions, but who do have access to the
app.