Add link to GOV.UK Homepage taxon from Edit Taxonomy page
The only way to navigate to this taxon page before was to choose a
random taxon in the list (hoping to choose one that does exist within
the Topic Taxonomy) and then following the link in the tree
visualisation.

Now we've added a link to this taxon page so that it is more easily
accessible.