Update schema.rb formatting
Rails 5.1.0 updated the default column type/argument spacing in the
schema dump. Before then there was an option to disable the column
formatting (introduced in [1]), but the default remained the same.

However, in 5.1.0, that formatting default changed in [2]. We upgraded
from Rails 5.0.2 to 5.1.4 on 2017-11-02, but there hasn't been any
additions to the schema since 2017-10-19.

This commit is separate to the work that I am doing to introduce a
Versions table, to avoid additional changes in the schema which doesn't
introduce any _real_ change.

[1] https://github.com/rails/rails/pull/25675
[2] https://github.com/rails/rails/commit/df84e9867219e9311aef6f4efd5dd9ec675bee5c