Change titles of world location taxons
The titles we have are of the form `<location name> and the UK` but these
are the same titles as those of the world location news pages for those
locations.  It's hard to tell the difference so we want to rename the
taxons to `UK help and services in <location name>`.  This one-off rake
task does that, and also changes the internal name too so that what you
see in content-tagger matches up with what you see in the live site.

The list of country slugs and names in the rake task was obtained with
the following code in the whitehall rails console:

    Hash[WorldLocation.order(:slug).where(world_location_type_id: 1).map {|w| [w.slug, w.name] }]

There are a handful of entries that may not be live taxons, but the rake
task will report these for us to check.