Save the changes of a Taxon when created, updated or restored
To record the changes to a Taxon, we rely on calling out to the
publishing API before the update request is made. Using the "current"
version from the API and the "updated" version from the validated form
parameters, a change diff is generated and stored in the Versions table.

This approach was chosen as it allows us to record changes over time,
regardless of its publishing state. For taxons which are published, we
would be able to harness the Publishing API edition versioning, and step
back in time with the user_facing_version query parameter.