Avoid recording nil -> [''] value for associated_taxons
By default, a hidden input field with a blank value is inserted before
the select tag. This is typically used for associated model
mass-assignment, but in our case, introduces an unnecessary change diff
of `nil -> ['']` on every form submission. We can use the
`include_hidden` option to diable the generation of this hidden input
field.