Display errors related to item in 'related content items' field
When JS is enabled, these input fields are replaced by a select2 UI
element. However, in using this library, the original Rails input fields
are hidden, including any related error messages that might prevent the
form from saving.

This happens to be the case in a recent Zendesk ticket [1] where the
user was unable to edit the fields, but also unable to self-diagnose and
resolve the problem because no error was being displayed, other than
"This form contains errors. Please correct them and try again."

In this commit, we include the error message related to the field, next
to the select2 item to allow the user updating the resource to rectify
the problem.

[1] https://govuk.zendesk.com/agent/tickets/2520366