Add 'bulk update phase' button to taxon page
A button within the phase/state panel has been added to 'Bulk update
phase', which will trigger a series of Sidekiq workers to update the
phase to the selected value on the confirmation page.

This currently works as expected for draft taxons, but there is still
work to be done to 'update and re-publish' taxons that are in the
published state.