Allow for the base paths to be directly edited.
We've removed the path auto-prefixing and changed the field from just
the path slug segment, to the whole base path. The user is free (to a
certain extent) to define the full base path, with some guidance on what
the prefix should be, based on what parent taxon is selected.