Add field to link to legacy taxons
The legacy taxon field allows users to add linked documents from the
base path. We perform a lookup by base path to verify the paths exist
(also verified by JS when adding a path) but to also convert the base
paths to content ids so that the linkset can be saved.