Remove sortable handler from path fields
Currently, only the legacy taxon field is using this path lookup field,
but the sortable component is redundant as the order of these paths are
not important.

Originally the field was introduced and planned to be used for the
related content items, where the order is important. However, this field
is currently not used for that input interface.