Make the user UID available on the tagging history page
Currently, the user details only show if that user has signed in to
Content Tagger. This should be fixable in the future by fetching the
data from Signon, but for now, just put the UID in the HTML as a
comment.

This should make it easier to find who this is in Signon.