Add facet values to tagging interface
This appends a facets field to the tagging interface and makes
all facet values in (currently a fixed group) appear as options
within the parent facet optgroup.

Facet values are saved as `facet_values` links similar to topics.

At the moment we only have one published facet group so this commit
includes a constant containing the content_id of the facet group, this
saves us making a lookup in the Publishing API for groups. We need
to modify the UI to enable choosing between multiple facet groups
when tagging content with facet values.