Use new filter key for business_activity
We need to send a different filter key to rummager
for the `AND` part of the business finder to work correctly.

The `and_facet_values` filter is identical to the `facet_values`
filter, but being a different field allows the default
behaviour (AND between fields OR within them) to work correctly.