Add fields to enable notifying when tagging facets
Facet tagging takes place outside the normal draft-to-publish
workflow. This means that users subscribed to a finder linked
to a facet group won't be notified when new items are tagged
to that facet group. This commit adds the ability to send a
notification with a custom message notifying subscribers of the
change.