Patch link to finder when tagging to facets
This supports linking back to a finder from a content page. To do
this the content item needs to be aware of which finders it can be
surfaced on.

This is being limited to finders that filter by linked based facets.
At the moment there is only one finder that does this, the business
readiness finder. Until that changes the finder content id is being
hard-coded.