Add test for creating related link to business finder
Tests that when tagging content items to the business readiness
finder, a related link will be added to the content item which
links back to the finder (this will always be the first item).

Also tests that the related link is removed when the content item
is no longer tagged to any facets from the finder, but that any
other related links are retained.