Create a related link when tagging to the business finder
When content is tagged to the business readiness finder, a
related link will be created in the content item.  This related
link will always be the first item.

The related link is removed when the content item is no longer
tagged to any of the facets within the finder.  All other
related links are retained.