Remove ability to pin to facets
The EU Exit Business Readiness Finder was the only finder using
pinning (this is where a content item was promoted to the top of
a facet by adding it as a related item on the content item of the
finder itself).

After retagging content, it has been decided that pinning is no
longer required, therefore this commit removes the ability to
pin content items when tagging to facets.