Add rake task to retag content for business finder
A number of business sectors in the EU Exit Business Readiness Finder
have been merged or split.  We therefore need to retag content that
is currently tagged to these sectors (through `facet_values`).

These rake tasks deal with the three scenarios.  One deals with the
replacement of one sector with another, another task retags a
specified list of content with their new sector and the third
patches content item links to remove deleted facet_values.

Full details of the changes:
https://docs.google.com/spreadsheets/d/1DDsWQvq3XoTwi3R3YS_z72ixON3mUl6aWjjsI2V4NHg/edit#gid=0